backend/requirements.txt
fastapi>=0.128.0
fastapi-users[beanie]>=15.0.3
beanie>=2.0.1
motor>=3.7.1
python-multipart>=0.0.6
python-jose[cryptography]>=3.3.0
passlib[bcrypt]>=1.7.4
uvicorn[standard]>=0.24.0
pydantic>=2.12.5
pydantic-settings>=2.12.0
email-validator>=2.0.0
pymongo>=4.16.0
bcrypt>=3.2.0
httpx>=0.25.0
PyPDF2>=3.0.1
Pillow>=12.1.0
reportlab>=4.0.7
openpyxl>=3.1.2
pytest>=7.4.3
